refactor of serializer validation into a mixin
